+++ Dyna/Charges/SwiftCharge.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 
 namespace Shockah.Dyna;
 
-public sealed class SwiftCharge : BaseDynaCharge, IRegisterable
+public sealed class SwiftCharge() : BaseDynaCharge($"{ModEntry.Instance.Package.Manifest.UniqueName}::SwiftCharge"), IRegisterable
 {
 	private static ISpriteEntry Sprite = null!;
 	private static ISpriteEntry LightsSprite = null!;
@@ -14,10 +14,6 @@ public sealed class SwiftCharge : BaseDynaCharge, IRegisterable
 	{
 		Sprite = ModEntry.Instance.Helper.Content.Sprites.RegisterSprite(ModEntry.Instance.Package.PackageRoot.GetRelativeFile("assets/Charges/Swift.png"));
 		LightsSprite = ModEntry.Instance.Helper.Content.Sprites.RegisterSprite(ModEntry.Instance.Package.PackageRoot.GetRelativeFile("assets/Charges/SwiftLight.png"));
-	}
-
-	public SwiftCharge() : base($"{ModEntry.Instance.Package.Manifest.UniqueName}::SwiftCharge")
-	{
 	}
 
 	public override Spr GetIcon(State state)
@@ -48,10 +44,7 @@ public sealed class SwiftCharge : BaseDynaCharge, IRegisterable
 			return;
 		var worldX = ship.x + partIndex;
 
-		combat.QueueImmediate(new ADrawCard
-		{
-			count = 3
-		});
+		combat.QueueImmediate(new ADrawCard { count = 3 });
 
 		var damageDone = new DamageDone { hitHull = true };
 		var raycastResult = new RaycastResult